Refactor the resulting html
The <head> section of the generated html is a mess, and extracting both
the metadata and style sheet out of it makes it a lot better. Tidy is
also run in the command line to format the html.

Now, this is still a little manual, but it doesn't take me more than a
minute. If it becomes annoying I'll just make 'Yet Another SSG.'